:ok_hand: Rename file and export the function by default from it
